Spring Framework Web Store. HomeWork lesson10-11 (new microservice statistics and recommendations, new front page, registrations new users, BigDecimal)

+++ core-service/src/main/java/com/geekbrains/spring/web/core/dto/OrderDto.java
@@ -1,6 +1,8 @@
 package com.geekbrains.spring.web.core.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -8,7 +10,7 @@ public class OrderDto {
     private Long id;
     private String username;
     private List<OrderItemDto> items;
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
     private String address;
     private String phone;
 }
