+++ core-service/src/main/java/com/geekbrains/spring/web/core/dto/OrderItemDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,6 +13,6 @@ public class OrderItemDto {
     private Long id;
     private String productTitle;
     private Integer quantity;
-    private Integer pricePerProduct;
-    private Integer price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 }
