+++ core-service/src/main/java/com/geekbrains/spring/web/core/repositories/specifications/ProductsSpecifications.java
@@ -3,12 +3,14 @@ package com.geekbrains.spring.web.core.repositories.specifications;
 import com.geekbrains.spring.web.core.entities.Product;
 import org.springframework.data.jpa.domain.Specification;
 
+import java.math.BigDecimal;
+
 public class ProductsSpecifications {
-    public static Specification<Product> priceGreaterOrEqualsThan(Integer price) {
+    public static Specification<Product> priceGreaterOrEqualsThan(BigDecimal price) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.greaterThanOrEqualTo(root.get("price"), price);
     }
 
-    public static Specification<Product> priceLessThanOrEqualsThan(Integer price) {
+    public static Specification<Product> priceLessThanOrEqualsThan(BigDecimal price) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.lessThanOrEqualTo(root.get("price"), price);
     }
