+++ core-service/src/main/java/com/geekbrains/spring/web/core/services/ProductsService.java
@@ -13,6 +13,7 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Service
@@ -20,7 +21,7 @@ import java.util.Optional;
 public class ProductsService {
     private final ProductsRepository productsRepository;
     private final ProductConverter productConverter;
-    public Page<Product> findAll(Integer minPrice, Integer maxPrice, String partTitle, Integer page) {
+    public Page<Product> findAll(BigDecimal minPrice, BigDecimal maxPrice, String partTitle, Integer page) {
         Specification<Product> spec = Specification.where(null);
         if (minPrice != null) {
             spec = spec.and(ProductsSpecifications.priceGreaterOrEqualsThan(minPrice));
